+++ CHANGELOG.md
@@ -5,9 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.8] - 2022-10-21
+### Added
+- `WithIgnoreFunc` `Option` which allows to ignore the error (return `nil`) under the specified condition.
+- `WithIgnoreString` `Option` which allows to ignore the error (return `nil`) if message contains the specified **string**.
+- `NewHTTPError` test.
+
 ## [1.0.7] - 2022-10-20
 ### Added
-- NewHTTPError for simple HTTP error messages, e.g.: "not found".
+- `NewHTTPError` for simple HTTP error messages, e.g.: "not found".
 
 ## [1.0.6] - 2022-10-04
 ### Changed